news_api.py
"""
Financial News API Integration

Handles fetching and processing financial news from various sources.
Uses NewsAPI and web scraping for comprehensive news coverage.
"""

import requests
from typing import Dict, List, Optional, Any
from datetime import datetime, timedelta
import logging
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class FinancialNewsAPI:
    """Wrapper for financial news API calls and web scraping"""
    
    def __init__(self, news_api_key: Optional[str] = None):
        """
        Initialize Financial News API client
        
        Args:
            news_api_key: NewsAPI key (optional, free tier available)
        """
        self.news_api_key = news_api_key
        self.news_api_base = "https://newsapi.org/v2"
        
        # Financial news sources
        self.finance_sources = [
            'bloomberg', 'cnbc', 'financial-times', 'reuters',
            'the-wall-street-journal', 'business-insider'
        ]
        
        logger.info("FinancialNewsAPI initialized")
    
    def get_market_news(self, category: str = 'general', limit: int = 10) -> List[Dict[str, Any]]:
        """
        Get latest financial market news
        
        Args:
            category: News category (general, stocks, crypto, economy)
            limit: Number of articles to return
            
        Returns:
            List of news articles
        """
        logger.info(f"Fetching {category} market news (limit: {limit})")
        
        if not self.news_api_key:
            logger.warning("NewsAPI key not provided, returning sample data")
            return self._get_sample_news(category, limit)
        
        try:
            # Define search queries based on category
            queries = {
                'general': 'stock market OR financial markets',
                'stocks': 'stock market OR equities OR shares',
                'crypto': 'cryptocurrency OR bitcoin OR ethereum',
                'economy': 'economy OR inflation OR interest rates'
            }
            
            query = queries.get(category, queries['general'])
            
            url = f"{self.news_api_base}/everything"
            params = {
                'apiKey': self.news_api_key,
                'q': query,
                'language': 'en',
                'sortBy': 'publishedAt',
                'pageSize': limit,
                'domains': 'bloomberg.com,cnbc.com,reuters.com,wsj.com,marketwatch.com'
            }
            
            response = requests.get(url, params=params, timeout=10)
            
            if response.status_code == 200:
                data = response.json()
                articles = []
                
                for article in data.get('articles', [])[:limit]:
                    articles.append({
                        'title': article.get('title', 'N/A'),
                        'description': article.get('description', 'N/A'),
                        'source': article.get('source', {}).get('name', 'N/A'),
                        'url': article.get('url', ''),
                        'published_at': article.get('publishedAt', ''),
                        'category': category
                    })
                
                logger.info(f"✅ Successfully fetched {len(articles)} news articles")
                return articles
            else:
                logger.error(f"❌ NewsAPI error: {response.status_code}")
                return self._get_sample_news(category, limit)
                
        except Exception as e:
            logger.error(f"❌ Error fetching news: {e}")
            return self._get_sample_news(category, limit)
    
    def get_stock_news(self, symbol: str, limit: int = 5) -> List[Dict[str, Any]]:
        """
        Get news articles related to a specific stock
        
        Args:
            symbol: Stock ticker symbol
            limit: Number of articles to return
            
        Returns:
            List of news articles about the stock
        """
        logger.info(f"Fetching news for {symbol}")
        
        if not self.news_api_key:
            return self._get_sample_stock_news(symbol, limit)
        
        try:
            url = f"{self.news_api_base}/everything"
            params = {
                'apiKey': self.news_api_key,
                'q': symbol,
                'language': 'en',
                'sortBy': 'publishedAt',
                'pageSize': limit,
                'domains': 'bloomberg.com,cnbc.com,reuters.com,wsj.com,marketwatch.com'
            }
            
            response = requests.get(url, params=params, timeout=10)
            
            if response.status_code == 200:
                data = response.json()
                articles = []
                
                for article in data.get('articles', [])[:limit]:
                    articles.append({
                        'title': article.get('title', 'N/A'),
                        'description': article.get('description', 'N/A'),
                        'source': article.get('source', {}).get('name', 'N/A'),
                        'url': article.get('url', ''),
                        'published_at': article.get('publishedAt', ''),
                        'symbol': symbol.upper()
                    })
                
                logger.info(f"✅ Successfully fetched {len(articles)} articles for {symbol}")
                return articles
            else:
                return self._get_sample_stock_news(symbol, limit)
                
        except Exception as e:
            logger.error(f"❌ Error fetching stock news: {e}")
            return self._get_sample_stock_news(symbol, limit)
    
    def scrape_market_sentiment(self) -> Dict[str, Any]:
        """
        Scrape market sentiment indicators from financial websites
        
        Returns:
            Dictionary with sentiment data
        """
        logger.info("Scraping market sentiment")
        
        # This is a simplified version - in production, implement proper scraping
        # with respect to robots.txt and rate limiting
        
        sentiment_data = {
            'fear_greed_index': 'N/A',
            'market_mood': 'Neutral',
            'vix_level': 'N/A',
            'put_call_ratio': 'N/A',
            'timestamp': datetime.now().strftime('%Y-%m-%d %H:%M:%S')
        }
        
        try:
            # Example: Could scrape CNN Fear & Greed Index
            # For now, return placeholder data
            logger.info("✅ Market sentiment data prepared")
            return sentiment_data
            
        except Exception as e:
            logger.error(f"❌ Error scraping sentiment: {e}")
            return sentiment_data
    
    def _get_sample_news(self, category: str, limit: int) -> List[Dict[str, Any]]:
        """Return sample news data when API is unavailable"""
        sample_articles = [
            {
                'title': 'Stock Market Reaches New Highs Amid Economic Recovery',
                'description': 'Major indices continue upward trend as investors remain optimistic about economic outlook.',
                'source': 'Sample Financial News',
                'url': 'https://example.com/article1',
                'published_at': datetime.now().isoformat(),
                'category': category
            },
            {
                'title': 'Fed Signals Potential Interest Rate Changes',
                'description': 'Central bank officials discuss monetary policy adjustments in response to inflation data.',
                'source': 'Sample Economic News',
                'url': 'https://example.com/article2',
                'published_at': (datetime.now() - timedelta(hours=2)).isoformat(),
                'category': category
            },
            {
                'title': 'Tech Sector Leads Market Performance',
                'description': 'Technology stocks drive gains as earnings reports exceed expectations.',
                'source': 'Sample Market Watch',
                'url': 'https://example.com/article3',
                'published_at': (datetime.now() - timedelta(hours=5)).isoformat(),
                'category': category
            }
        ]
        
        logger.info(f"⚠️  Returning {limit} sample news articles")
        return sample_articles[:limit]
    
    def _get_sample_stock_news(self, symbol: str, limit: int) -> List[Dict[str, Any]]:
        """Return sample stock news when API is unavailable"""
        sample_articles = [
            {
                'title': f'{symbol} Reports Strong Quarterly Earnings',
                'description': f'{symbol} exceeds analyst expectations with robust revenue growth.',
                'source': 'Sample Financial News',
                'url': 'https://example.com/stock1',
                'published_at': datetime.now().isoformat(),
                'symbol': symbol.upper()
            },
            {
                'title': f'Analysts Upgrade {symbol} Price Target',
                'description': f'Wall Street analysts raise their outlook for {symbol} following positive developments.',
                'source': 'Sample Market Analysis',
                'url': 'https://example.com/stock2',
                'published_at': (datetime.now() - timedelta(hours=3)).isoformat(),
                'symbol': symbol.upper()
            }
        ]
        
        logger.info(f"⚠️  Returning {limit} sample articles for {symbol}")
        return sample_articles[:limit]
    
    def get_economic_calendar(self, days: int = 7) -> List[Dict[str, Any]]:
        """
        Get upcoming economic events and data releases
        
        Args:
            days: Number of days ahead to look
            
        Returns:
            List of economic events
        """
        logger.info(f"Fetching economic calendar for next {days} days")
        
        # Sample economic events - in production, integrate with economic calendar API
        sample_events = [
            {
                'date': (datetime.now() + timedelta(days=1)).strftime('%Y-%m-%d'),
                'event': 'Federal Reserve Interest Rate Decision',
                'importance': 'High',
                'forecast': 'No change expected'
            },
            {
                'date': (datetime.now() + timedelta(days=3)).strftime('%Y-%m-%d'),
                'event': 'Monthly Employment Report',
                'importance': 'High',
                'forecast': 'Job growth expected'
            },
            {
                'date': (datetime.now() + timedelta(days=5)).strftime('%Y-%m-%d'),
                'event': 'Consumer Price Index (CPI)',
                'importance': 'High',
                'forecast': 'Inflation data release'
            }
        ]
        
        logger.info(f"✅ Returning {len(sample_events)} economic events")
        return sample_events
